refactor: move inference test helpers into test module
